Made Sentry model and Player Target function
Made a formation of meshes and materials to resemble a sentry gun one would find in CoD. Gave said sentry its own unique script to always target. Gave 'The Player' a TargetableThing script to be able to be targeted by the sentry. Had to modify PlayerTargeting to have a min distance as to not target themselves.